Make both ternary branches explicitly int64_t
The int32 branch now casts to int64_t rather than relying on the implicit
conversion the ternary's common type performs, so both branches have the
same type at the point of selection.

The cast sits outside SubtleMustCopy rather than inside it. That ordering
is what keeps the earlier review fix intact: SubtleMustCopy takes a
const T& and reads through a volatile pointer to the address it binds, so
it has to bind to the tensor buffer to be worth anything. Casting first
would hand it a temporary and quietly restore the problem. Casting its
result is free -- the volatile read has already happened on the buffer,
and widening a copied int32 cannot lose information.

No behavioural change: the ternary already yielded int64_t by promotion,
so this only makes the conversion explicit and its position deliberate.
Error strings are untouched and the three regression tests still assert
the same messages.

Also corrects a docstring that still described the bound as 2^31; it has
been kint32max since 12ad001.

+++ tensorflow/core/kernels/segment_reduction_ops_impl.h
@@ -523,7 +523,8 @@ class UnsortedSegmentReductionOp : public OpKernel {
     // subsequently used (see SubtleMustCopy).
     const int64_t num_segments_val =
         num_segments.dtype() == DT_INT32
-            ? internal::SubtleMustCopy(num_segments.scalar<int32_t>()())
+            ? static_cast<int64_t>(
+                  internal::SubtleMustCopy(num_segments.scalar<int32_t>()()))
             : internal::SubtleMustCopy(num_segments.scalar<int64_t>()());
     OP_REQUIRES(context, num_segments_val >= 0,
                 absl::InvalidArgumentError(
